+++ module-api/src/main/java/com/welcommu/moduleapi/logging/AuditLogController.java
@@ -1,9 +1,7 @@
 package com.welcommu.moduleapi.logging;
 
-import com.welcommu.moduledomain.logging.AuditLog;
 import com.welcommu.moduledomain.logging.enums.ActionType;
 import com.welcommu.moduledomain.logging.enums.TargetType;
-import com.welcommu.moduledomain.project.Project;
 import com.welcommu.moduleservice.logging.AuditLogSearchService;
 import com.welcommu.moduleservice.logging.dto.AuditLogResponse;
 import com.welcommu.moduleservice.logging.dto.LogsWithCursor;
@@ -11,14 +9,11 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
